% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_related_functions.R
\name{add_outgroup}
\alias{add_outgroup}
\title{add_outgroup}
\usage{
add_outgroup(subtypes, path_output, LANL_seq, output_names,
  outgroup_size = 5, index = NA, verbose = TRUE)
}
\arguments{
\item{subtypes}{XXXX}

\item{path_output}{XXXX}

\item{LANL_seq}{XXXX}

\item{output_names}{XXXX#'}

\item{outgroup_size}{XXXX}

\item{index}{XXXX}

\item{verbose}{XXXX}
}
\value{
dqsfds
}
\description{
Script for:
}
