% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_related_functions.R
\name{align_seq}
\alias{align_seq}
\title{align_seq}
\usage{
align_seq(names_file, aligner, thread = 5, index = NA,
  verbose = TRUE)
}
\arguments{
\item{names_file}{XXXX}

\item{aligner}{XXXX}

\item{thread}{XXXX}

\item{index}{XXXX}

\item{verbose}{XXXX}
}
\value{
dqsfds
}
\description{
Script for:
}
