% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_related_functions.R
\name{blastn}
\alias{blastn}
\title{blastn}
\usage{
blastn(names_seq, names_db, nbre_close = 20, index = NA,
  verbose = TRUE)
}
\arguments{
\item{names_seq}{XXXX}

\item{names_db}{XXXX}

\item{nbre_close}{XXXX}

\item{index}{XXX}

\item{verbose}{XXXX}
}
\value{
dqsfds
}
\description{
Script for:
}
