% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building_tree_functions.R
\name{building_tree}
\alias{building_tree}
\title{building_tree}
\usage{
building_tree(names_file_aligned_ndrm, path_output,
  tree_builder = "fasttree", bs.n = 50, bs.id = NA, verbose = TRUE)
}
\arguments{
\item{names_file_aligned_ndrm}{XXXX}

\item{path_output}{XXXX}

\item{tree_builder}{XXXX}

\item{bs.n}{XXXX}

\item{bs.id}{XXXX}

\item{verbose}{XXXX}
}
\value{
dqsfds
}
\description{
Script for:
}
