% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_related_functions.R
\name{check_HXB2}
\alias{check_HXB2}
\title{check_HXB2}
\usage{
check_HXB2(name_file, name_HXB2_file, index = NA, verbose = TRUE)
}
\arguments{
\item{name_file}{XXXX}

\item{name_HXB2_file}{XXXX}

\item{index}{XXXX}

\item{verbose}{XXXX}
}
\value{
dqsfds
}
\description{
Script for:
}
