% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clades_related_functions.R
\name{clades_extracting}
\alias{clades_extracting}
\title{clades_extracting}
\usage{
clades_extracting(renamed_tree_names, path_colored_tree, path_output,
  focus_group, bs.n = 50, bs.id = NA, verbose = T)
}
\arguments{
\item{renamed_tree_names}{XXXX}

\item{path_colored_tree}{XXXX}

\item{path_output}{XXXX}

\item{focus_group}{XXXX}

\item{bs.n}{XXXX}

\item{bs.id}{XXXX}

\item{verbose}{XXX}
}
\value{
dqsfds
}
\description{
Script for:
}
