% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{clean_LANL}
\alias{clean_LANL}
\title{Clean the LANL database.}
\usage{
clean_LANL(name_file, info_number, verbose = TRUE)
}
\arguments{
\item{name_file}{a character vector, the LANL file names by subtype.}

\item{info_number}{an integer, the number of information in the sequence names.}

\item{verbose}{Write stuff.}
}
\description{
Remove sequences which are duplicates (with regards to accesion
  number or name) and remove sequences for which the name does not contain
  the required information.
}
