% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{extract_HXB2}
\alias{extract_HXB2}
\title{extract_HXB2}
\usage{
extract_HXB2(path_LANL_sequences, prefixe_LANL, name_HXB2, name_HXB2_file,
  verbose = TRUE)
}
\arguments{
\item{path_LANL_sequences}{character string giving the path to the LANL datasets.}

\item{prefixe_LANL}{character string giving the first characters of the LANL datasets.}

\item{name_HXB2}{character string giving the name of the HXB2 sequence.}

\item{name_HXB2_file}{character string giving the path name of the output.}

\item{verbose}{Write stuff. [default = TRUE]}
}
\value{
dsfdsf
}
\description{
Extract the HXB2 sequence from a LANL dataset.
}
