% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_related_functions.R
\name{find_identical_sequences}
\alias{find_identical_sequences}
\title{find_identical_sequences}
\usage{
find_identical_sequences(names_file_aligned, path_sequences_meta,
  dna_region, index = NA, verbose = TRUE)
}
\arguments{
\item{names_file_aligned}{XXXX}

\item{path_sequences_meta}{XXXX}

\item{dna_region}{XXXX}

\item{index}{XXXX}

\item{verbose}{XXXX}
}
\value{
dqsfds
}
\description{
Script for:
}
