% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_related_functions.R
\name{get_closest_seq}
\alias{get_closest_seq}
\title{get_closest_seq}
\usage{
get_closest_seq(LANL_seq, seq, path_output, output_names, max_closest,
  nbre_close = NULL, min_percent, name_HXB2_file = NULL, index = NA,
  verbose = TRUE)
}
\arguments{
\item{LANL_seq}{XXXX}

\item{seq}{XXXX}

\item{path_output}{XXXX}

\item{output_names}{XXXX}

\item{max_closest}{XXXX}

\item{nbre_close}{XXXX}

\item{min_percent}{XXXX}

\item{name_HXB2_file}{XXXX}

\item{index}{XXXX}

\item{verbose}{XXXX}
}
\value{
dqsfds
}
\description{
Script for:
}
