% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_related_functions.R
\name{makeblastdb}
\alias{makeblastdb}
\title{makeblastdb}
\usage{
makeblastdb(seq_file, output_file, index = NA, verbose = T)
}
\arguments{
\item{seq_file}{XXXX}

\item{output_file}{XXXX}

\item{index}{XXX}

\item{verbose}{XXXX}
}
\value{
dqsfds
}
\description{
Script for:
}
