% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clades_related_functions.R
\name{MRCA_correction}
\alias{MRCA_correction}
\title{clades_extracting}
\usage{
MRCA_correction(tree, traits, MRCAs, state)
}
\arguments{
\item{tree}{XXXX}

\item{traits}{XXXX}

\item{MRCAs}{XXXX}

\item{state}{XXXX}
}
\value{
dqsfds
}
\description{
Script for:
}
