% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building_tree_functions.R
\name{postprocess_tree}
\alias{postprocess_tree}
\title{postprocess_tree}
\usage{
postprocess_tree(tree_names, path_outgroup, name_HXB2, bs.n = 50,
  bs.id = NA, verbose = TRUE)
}
\arguments{
\item{tree_names}{XXXX}

\item{path_outgroup}{XXXX}

\item{name_HXB2}{XXXX}

\item{bs.n}{XXXX}

\item{bs.id}{XXXX}

\item{verbose}{XXXX}
}
\value{
dqsfds
}
\description{
Script for:
}
