% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_related_functions.R
\name{rm_DRM}
\alias{rm_DRM}
\title{rm_DRM}
\usage{
rm_DRM(names_file_aligned, name_HXB2, index = NA, verbose = TRUE)
}
\arguments{
\item{names_file_aligned}{XXXX}

\item{name_HXB2}{XXXX}

\item{index}{XXXX}

\item{verbose}{XXXX}
}
\value{
dqsfds
}
\description{
Script for:
}
