% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{rm_gaps_LANL}
\alias{rm_gaps_LANL}
\title{rm_gaps_LANL}
\usage{
rm_gaps_LANL(LANL_seq, code.from = "-", code.to = "n",
  verbose = TRUE)
}
\arguments{
\item{XXX}{XXXX}
}
\description{
Script for:
\describe{
  \item{\code{XXX}}{XXX}
  }
}
