% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_related_functions.R
\name{seq_rm_drugresistance2}
\alias{seq_rm_drugresistance2}
\title{seq_rm_drugresistance2}
\usage{
seq_rm_drugresistance2(seq, outfile = NA, name_HXB2)
}
\arguments{
\item{seq}{XXXX}

\item{outfile}{XXXX}

\item{name_HXB2}{XXXX}
}
\value{
dqsfds
}
\description{
Script for:
}
