% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clades_related_functions.R
\name{tips_relabelling.internal}
\alias{tips_relabelling.internal}
\title{tips_relabelling.internal}
\usage{
tips_relabelling.internal(tree, person, sequences_meta, vars_for_label,
  local_indicator = "PR/RT")
}
\arguments{
\item{tree}{XXXX}

\item{person}{XXXX}

\item{sequences_meta}{XXX}

\item{vars_for_label}{XXXX}

\item{local_indicator}{XXXX}
}
\value{
dqsfds
}
\description{
Script for:
}
