% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clades_related_functions.R
\name{tips_relabelling}
\alias{tips_relabelling}
\title{tips_relabelling}
\usage{
tips_relabelling(tree_names, person, sequences_meta, Country_db,
  vars_for_label, local_group = "WA", focus_group = "KC",
  focus_subgroup = NULL, index_subgroup = NULL, name_subgroup = NULL,
  negative = "non", local_indicator = "PR/RT",
  focus_indicator = TRUE, separator = "___", blacklist_paths = NULL,
  blacklist_names = NULL, Unknown = "Unknown", local_tip_split = "-",
  external_tip_split = "[.]", local_index_location = 8,
  external_index_location = 2, bs.n = 50, bs.id = NA,
  verbose = TRUE)
}
\arguments{
\item{tree_names}{XXXX}

\item{person}{XXXX}

\item{sequences_meta}{XXXX}

\item{Country_db}{XXXX}

\item{vars_for_label}{XXXX}

\item{local_group}{XXX}

\item{focus_group}{XXXX}

\item{focus_subgroup}{XXX}

\item{index_subgroup}{XXXX}

\item{name_subgroup}{XXX}

\item{negative}{XXXX}

\item{local_indicator}{XXX}

\item{focus_indicator}{XXXX}

\item{separator}{XXX}

\item{blacklist_paths}{XXXX}

\item{blacklist_names}{XXX}

\item{Unknown}{XXXX}

\item{local_tip_split}{XXX}

\item{external_tip_split}{XXXX}

\item{local_index_location}{XXX}

\item{external_index_location}{XXXX}

\item{bs.n}{XXXX}

\item{bs.id}{XXXX}

\item{verbose}{XXXX}
}
\value{
dqsfds
}
\description{
Script for:
}
