% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clades_related_functions.R
\name{wrap_extract.clades}
\alias{wrap_extract.clades}
\title{extract.clades_wrap}
\usage{
wrap_extract.clades(tree, traits, MRCAs, focus_group)
}
\arguments{
\item{tree}{XXXX}

\item{traits}{XXXX}

\item{MRCAs}{XXXX}

\item{focus_group}{XXXX}
}
\value{
dqsfds
}
\description{
Script for:
}
