% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clades_related_functions.R
\name{wrap_phyloscanner_analyse_tree}
\alias{wrap_phyloscanner_analyse_tree}
\title{wrap_phyloscanner_analyse_tree}
\usage{
wrap_phyloscanner_analyse_tree(tree_names, path_output,
  opt_regex = "([[:alpha:]]+)___.*", bs.n = 50, bs.id = NA,
  verbose = TRUE)
}
\arguments{
\item{tree_names}{XXXX}

\item{path_output}{XXXX}

\item{opt_regex}{XXXX}

\item{bs.n}{XXXX}

\item{bs.id}{XXXX}

\item{verbose}{XXX}
}
\value{
dqsfds
}
\description{
Script for:
}
