% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{write_option_file}
\alias{write_option_file}
\title{write_option_file}
\usage{
write_option_file(output_file = "./options.json",
  path_RDS = "./extdata/FakeData/",
  RDS_names = c("sequences_meta_modified.rds", "person_modified.rds",
  "Country_db.rds"), path_LANL_seq = "./extdata/LANL/",
  LANL_seq_names = c("LANL_Subtype_01_AE_stripped.fasta",
  "LANL_Subtype_A1_stripped.fasta"), path_seq = "./extdata/Subtype/",
  seq_names = c("newseq_Subtype_01_AE.fasta", "newseq_Subtype_A1.fasta"),
  path_file_aligned = "./extdata/Closest/",
  names_file_aligned = c("newseq_LANL_Subtype_01_AE",
  "newseq_LANL_Subtype_A1"), path_tree = "./extdata/Tree/",
  tree_names = c("newseq_LANL_Subtype_01_AE_mafft_aligned_ndrm",
  "newseq_LANL_Subtype_A1_mafft_aligned_ndrm"),
  path_colored_tree = "./extdata/Colored_tree/",
  path_clades = "./extdata/Clades/", local_group = "WA",
  focus_group = "KC", bs.n = 2, subtypes = c("01_AE", "A1"),
  job_index = NA, verbose = "true")
}
\arguments{
\item{output_file}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}

\item{path_RDS}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}

\item{RDS_names}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}

\item{path_LANL_seq}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}

\item{LANL_seq_names}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}

\item{path_seq}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}

\item{seq_names}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}

\item{path_file_aligned}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}

\item{names_file_aligned}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}

\item{path_tree}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}

\item{tree_names}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}

\item{path_colored_tree}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}

\item{path_clades}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}

\item{local_group}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}

\item{focus_group}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}

\item{bs.n}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}

\item{subtypes}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}

\item{job_index}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}

\item{verbose}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}
}
\value{
dsfdsf
}
\description{
Extract the HXB2 sequence from a LANL dataset.
}
