% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{write_qsub}
\alias{write_qsub}
\title{write_qsub}
\usage{
write_qsub(Rscript, options_file = NA, path_output = "./",
  job_title = NA, hpc.q = "pqeelab", hpc.select = 1,
  hpc.walltime = 24, hpc.mem = "1gb", hpc.nproc = 1,
  hpc.load = big.phylo:::HPC.CX1.IMPERIAL.LOAD, hpc.array = NA,
  log = TRUE, multiple_log = TRUE)
}
\arguments{
\item{Rscript}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}

\item{options_file}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}

\item{path_output}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}

\item{job_title}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}

\item{hpc.q}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}

\item{hpc.select}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}

\item{hpc.walltime}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}

\item{hpc.mem}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}

\item{hpc.nproc}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}

\item{hpc.load}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}

\item{hpc.array}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}

\item{log}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}

\item{multiple_log}{XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx}
}
\value{
dsfdsf
}
\description{
Extract the HXB2 sequence from a LANL dataset.
}
