% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clades_related_functions.R
\name{phyloscanner_postprocess}
\alias{phyloscanner_postprocess}
\title{phyloscanner_postprocess}
\usage{
phyloscanner_postprocess(renamed_tree_names, path_output, focus_group,
  separator, bs.n = 50, bs.id = NA, verbose = TRUE)
}
\arguments{
\item{renamed_tree_names}{XXXX}

\item{path_output}{XXXX}

\item{focus_group}{XXXX}

\item{separator}{XXXX}

\item{bs.n}{XXXX}

\item{bs.id}{XXXX}

\item{verbose}{XXX}
}
\value{
dqsfds
}
\description{
Script for:
}
